Fix concurrent writing
Actually, it might be that all that effort was for nought: I've just
realized we still have sync issues since out write calls are not for
isolated chunks of stuff but often very many for one `.print` call.
Thus, we need to lock the target in `.print`.

We could probably do that off-thread by putting the item into a channel,
but that'd require the item to implement `Send`. I'm not sure I want to
do that (robustness principle and all that). I might still evaluate this
in a spike branch just because I don't have enough experience to judge
the effects of this.

This commit surely adds a bit of tech debt, as we `Output` is now
basically a `Vec<Arc<Mutex<Arc<InternalFormatter>>>`. Sorry.