Make targets Send and Sync
Human and JSON targets now spawn a thread that will do all the stuff.
This mean we now have to pass in an initializer closure so buffers don't
have to be send.

Oh, and since we are dealing with threads, I've also had the pleasure of
adding some channels here and there. We have one unbounded channel that
takes all write requests. We keep the sender of that one, and pass the
receiver to the thread. To get stuff out of the thread, we have another
channel. This one is more synchronous by having a capacity of 0 -- IIRC
that means it'll block until we have received its content. Of this
"response channel", we pass the sender into the thread and keep a single
receiver for ourselves.

While initializing the formatter, this is also the channel we block on
to see if the thread started successfully or if there was an error.
Because, yes, our initializer can return an error.

Along the way, I've also added a `flush` method, so that we can now in
tests just call that instead of dropping the output explicitly as a sync
point. (This too is synchronized using the response channel).