Refactor output struct to be an `Arc<Mutex<..>>`
This is required to be able to pass it off to a global log hook. I know
we now have a bunch of `Arc<Mutex<..>>` stuff in this codebase and I
need to clean it up. For now, it's not a blocking issue, though.

Additionally, this internal mutability switch trades in a few `mut`
usages for more `?`.