Add method to register `Output` as global logger
Thanks to the previous refactoring, we can now write

    log::set_boxed_logger(Box::new(self.clone()))?;

as well as implement the `Log` trait pretty easily.

I have added a simple LogMessage style, but ideally, the user would be
able to tweak this. It shouldn't be too hard -- I imagine supplying a
type `T: FromLogRecord` (or similar) to a `use_as_logger_with_format`
method.